long -> float for averages

+++ packages/kbn-telemetry-tools/src/tools/__fixture__/parsed_working_collector.ts
@@ -34,7 +34,7 @@ export const parsedWorkingCollector: ParsedUsageCollection = [
         },
         my_index_signature_prop: {
           avg: {
-            type: 'long',
+            type: 'float',
           },
           count: {
             type: 'long',
